b/pyproject.toml
[tool.poetry]
name = "back-taller1"
version = "0.1.0"
description = ""
authors = ["ju4nsepulveda <juan.sepulveda.z@ab-inbev.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.3"
fastapi = ">=0.101.0,<0.102.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlmodel = "^0.0.14"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
celery = "^5.3.6"
redis = "^5.0.1"
google = "^3.0.0"
google-cloud-storage = "^2.14.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
